requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.4.0
docker==7.1.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
requests==2.32.3
urllib3==2.2.3
yarl==1.13.1

# only needed when running on Windows
pywin32==306; platform_system == "Windows"
